+++ checkconfig.php
@@ -198,7 +198,18 @@ if (!str_ends_with($base, '/')) {
     $base .= '/';
 }
 echo 'Base URL detected by the script: ' . $base . '<br>';
-echo 'Full URL specified in $config[\'cops_full_url\']: ' . $config['cops_full_url'] . '<br><br>';
+echo 'Full URL specified in $config[\'cops_full_url\']: ' . $config['cops_full_url'] . '<br>';
+if (Route::hasTrustedProxies()) {
+    echo 'Trusted proxies configured: ' . $config['cops_trusted_proxies'] . '<br>';
+    echo 'Trusted headers configured: ' . json_encode($config['cops_trusted_headers']) . '<br>';
+    echo 'REMOTE_ADDR: ' . ($_SERVER['REMOTE_ADDR'] ?? '') . '<br>';
+    foreach ($config['cops_trusted_headers'] as $name) {
+        $header = 'HTTP_' . strtoupper(str_replace('-', '_', $name));
+        echo $header . ': ' . ($_SERVER[$header] ?? '') . '<br>';
+    }
+    echo 'Base URL via trusted proxies: ' . Route::base() . '<br>';
+}
+echo '<br>';
 echo 'SCRIPT_NAME: ' . ($_SERVER['SCRIPT_NAME'] ?? '') . '<br>';
 echo 'HTTP_HOST: ' . ($_SERVER['HTTP_HOST'] ?? '') . '<br>';
 echo 'SERVER_NAME: ' . ($_SERVER['SERVER_NAME'] ?? '') . '<br>';