+++ config_default.php
@@ -42,6 +42,16 @@ $config['calibre_internal_directory'] = '';
 $config['cops_full_url'] = '';
 
 /*
+ * As an alternative for cops_full_url above, if you're using a reverse proxy and you want to
+ * change how COPS is accessed depending on the entrypoint (e.g. direct, local network, internet)
+ * you can define trusted proxies and trusted headers here (dev only)
+ * Note: using symfony/http-foundation to support X-Forwarded-* and Forwarded headers from proxies
+ * @see https://symfony.com/doc/current/deployment/proxies.html
+ */
+$config['cops_trusted_proxies'] = '';
+$config['cops_trusted_headers'] = [];
+
+/*
  * Number of recent books to show
  */
 $config['cops_recentbooks_limit'] = '50';